Adds multi-part namespace support to catalogs
Enables catalogs to support hierarchical organization of tables through multi-part namespaces.

Introduces changes to the Catalog class and its implementations (DynamoDB, REST, SQL) to handle multi-part namespaces.
A new document is added to explain the usage of the multi-part namespace feature.

enable multipart namespace

Multipart namespace efers to a namespace that is structured with multiple levels or parts and default separated by a delmiiter dot (`.`)